move signature to sign protocol
