+++ protocols/sign/result.go
@@ -1,7 +1,5 @@
 package sign
 
-import "github.com/taurusgroup/cmp-ecdsa/protocols/sign/signature"
-
 type Result struct {
-	Signature *signature.Signature
+	Signature *Signature
 }
